requirements.txt
flask==3.0.3
pinecone==5.0.1
groq==0.9.0
httpx==0.27.2
pypdf==4.3.1
python-dotenv==1.0.1
gunicorn==22.0.0
werkzeug==3.0.3